+++ src/main/java/com/maciejwalkowiak/mercury/core/api/ApiController.java
@@ -1,6 +1,7 @@
 package com.maciejwalkowiak.mercury.core.api;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.hateoas.Link;
 import org.springframework.hateoas.ResourceSupport;
 import org.springframework.http.HttpEntity;
 import org.springframework.stereotype.Controller;
@@ -48,5 +49,19 @@ class ApiController {
 	}
 
 	private static class ApiResource extends ResourceSupport {
+	}
+
+	/**
+	 * Spring HATEOAS performs URL encoding and replaces characters "{" and "}" that are useful to show templated URL.
+	 *
+	 * BracketsLink is a hacky class that takes {@link org.springframework.hateoas.Link}
+	 * from Spring HATEOAS package and brings back brackets "{" and "}"
+	 *
+	 * @author Maciej Walkowiak
+	 */
+	private static class BracketsLink extends Link {
+		public BracketsLink(Link link) {
+			super(link.getHref().replaceAll("%7B", "{").replaceAll("%7D", "}"), link.getRel());
+		}
 	}
 }
