+++ src/main/java/com/maciejwalkowiak/mercury/core/message/MessageController.java
@@ -1,6 +1,5 @@
 package com.maciejwalkowiak.mercury.core.message;
 
-import com.fasterxml.jackson.annotation.JsonView;
 import com.maciejwalkowiak.mercury.core.api.HateoasController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.hateoas.Link;
