+++ src/main/java/com/maciejwalkowiak/mercury/mail/sendgrid/SendGridHealthIndicator.java
@@ -3,7 +3,6 @@ package com.maciejwalkowiak.mercury.mail.sendgrid;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.actuate.health.AbstractHealthIndicator;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.stereotype.Component;
