+++ src/test/java/com/maciejwalkowiak/mercury/core/message/MessengerTest.java
@@ -1,7 +1,6 @@
 package com.maciejwalkowiak.mercury.core.message;
 
 import com.maciejwalkowiak.mercury.mail.common.SendMailRequest;
-import com.maciejwalkowiak.mercury.mail.common.SendMailRequestBuilder;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
@@ -13,15 +12,15 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
-public class MessengerImplTest {
+public class MessengerTest {
 	@InjectMocks
-	private MessengerImpl messenger;
+	private Messenger messenger;
 	@Mock
 	private MessageService messageService;
 	@Mock
 	private MessageNotifier messageNotifier;
 
-	private final SendMailRequest request = new SendMailRequestBuilder()
+	private final SendMailRequest request = new SendMailRequest.Builder()
 			.to("foo@bar.com")
 			.subject("subject")
 			.text("content")
